python-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
httpx>=0.25.2
celery[redis]>=5.3.4
redis>=5.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1
